# back-end/tel_openapi.yaml
openapi: "3.0.2"
info:
  description: Control Center API description
  title: Control Center
  version: 1.0.0
paths:
...
  /reports:
    get:
      tags:
        - reports
      summary: Get all reports
      operationId: get_reports
      responses:
        '200':
          description: Successful retrieval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/reportArray"
        '404':
          description: Not Found
        '401':
          description: Not Authorized for this operation
        '500':
          description: Internal server error
        '403':
          description: Forbidden to use this endpoint
    post:
      tags:
        - reports
      summary: Close an incident
      description: This endpoint is used to close an incident, which results in reports opening for all users involved.
      operationId: post_reports
      requestBody:
        content:
          application/json:   
            schema:
              $ref: "#/components/schemas/report"
        required: true
      responses:
        '200':
          description: Successful submition
        '404':
          description: Not Found
        '500':
          description: Internal server error
        '401':
          description: Not Authorized for this operation 
        '400':
          description: Bad request
        '403':
          description: Forbidden to use this endpoint
  /reports/{reportId}:
    get:
      tags:
        - reportsId
      summary: Get a report
      operationId: get_reports_byid
      parameters:
        - name: reportId
          in: path
          description: ID of report
          required: true
          schema:
            $ref: "#/components/schemas/reportId"
      responses:
        '200':
          description: Successful retrieval
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/report"
        '404':
          description: Not Found
        '401':
          description: Not Authorized for this operation 
    delete:
      tags:
        - reportsId
      summary: Delete a report
      operationId: delete_reports_byid
      parameters:
        - name: reportId
          in: path
          description: ID of report
          required: true
          schema:
            $ref: "#/components/schemas/reportId"
      responses:
        '200':
          description: Successful deletion
        '404':
          description: Not Found
        '401':
          description: Not Authorized for this operation 
        '403':
          description: Forbidden to use this endpoint
    put:
      tags:
        - reportsId
      summary: Update a report
      operationId: put_reports_byid
      parameters:
        - name: reportId
          in: path
          description: ID of report
          required: true
          schema:
            $ref: "#/components/schemas/reportId"
      requestBody:
        content:
          application/json:    
            schema:
              $ref: "#/components/schemas/report"
      responses:
        '200':
          description: Successful update
        '404':
          description: Not Found
        '401':
          description: Not Authorized for this operation
        '403':
          description: Forbidden to use this endpoint
    patch:
      tags:
        - reportsId
      summary: Update a report
      description: This endpoint is specifically used by the author of the open to report to provide his comments. about the incident. When the last report closes, a formal report opens automatically.
      operationId: patch_reports_byid
      parameters:
        - name: reportId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/reportId"
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/comments"
      responses:
        '200':
          description: Successful update
        '404':
          description: Not Found
        '401':
          description: Not Authorized for this operation
        '400':
          description: Bad request
        '403':
          description: Forbidden to use this endpoint
        '500':
          description: Internal server error
...
components:
  schemas:
    agencyId:
      type: string
      description: String containing the agency id
    userId:
      type: string
      description: String containing the user id
    incidentId:
      type: string
      description: String containing the incident id
    reportId:
      type: string
      description: String containing the report id
...
    report:
      type: object
      required:
        - reportId
      properties:
        reportId:
          $ref: "#/components/schemas/reportId"
        title:
          type: string
          description: Contains the unique report title
        userId:
          $ref: "#/components/schemas/userId"
        incidentId:
          $ref: "#/components/schemas/incidentId"
        agencyId:
          $ref: "#/components/schemas/agencyId"
        date:
          type: string
          description: The start date of the initial incident
        comments:
          type: string
          description: String containing the body of the report
        isOpen:
          type: boolean
...
    reportArray:
      type: array
      items:
        $ref: "#/components/schemas/report"
...
    comments:
      type: string
      description: the user comments of the report
      example:
        comments: "blah blah blah"